Adding PlayerController, Charactermotor now base class for all humans
XD
