+++ Assets/StartMenu.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 3ae753b8f64654e0aa1912fab86ad8c0
+timeCreated: 1515098148
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
